NavGraph : Implemented a more complete version of the movement system.
The Edges and Nodes are now properly implemented into separate classes working together move the player arond the scene.

The EdgeTrigger implemented in previous trigger is nwo solely responsible for handling user input and starting the movement.
All SWS and spline movement related code has been moved to EdgeMovement, including setting up events at start and end of movement to notify the Nodes of movements.
The Node class is mostly about holding a position and events when the player enters or leaves the node. (These events are invoked by the start and end events of the EdgeMovements)

Also added a NavGraphManager, a singleton to keep track of the currently active Node and Object that is to be moved along the edges.
It might come with additional functionality in the near future, like initializing a complete NavGraph by itself.

Possible improvements :
+ Give the EdgeMovement the possibility to create/update the PathManager it holds from the start and end Nodes, plus a list of intermediate positions.
+ Give the Manager the possibility to create/update the NavGrap from a dedicated editor interface. (Inputs : Nodes positions, Nodes adjacency + intermediate positions, position of the EdgeTriggers relative to a Node)