Set up a basic underwater ambiance for level02
See the lighting editor for ambient and fog settings. (I recommend leaving the fog thickness at 0.038)
The ambient is set to a darker tone to make the shadows contrast with the light eerie blue from the directional light.
The fog also prevents the player form seeing the bottom of the pit or the edges of the map.

The skybox is generated from the standard procedural shader wich is made for actual skies, and not underwater environements.
We may want to make or find a more suitable version.

The resuly is pretty basic and unrealistic, but shows that we can achieve something pretty without particles or even textures.

Feel free to tinker with the colors for the fog, ambient, skybox and light.