First go at level03 lighting. Teh floor is lava !
The main source of lighting in this level is the lava pool, wich has an emissive material.

Properlely baking the light from the emissive surfaces was troublesome, since stripes/artifacts appeared in the lightmap.
Solved the issue by setting the imported models for the level to genereate their UVs.

The lightmap scale factors for the elements in the scene were heavily rebalanced to improve quality of light on smaller elements (plateforms).