Created a preload scene to initialize the app.
The purpose of this scene is to do every needed intialization before moving on to the actual game.
The scene contains a gameobject that will holds the state of the application through the scenes, and hold all general/transversal behaviour.
This game project requires three general behaviours to work :
- moving to another level
- loading/writting saves
- tracking players advancement : runes collected, activated objects in levels.
+ sound, and user user input could also be moved here.

More detailed explanation are given here :
http://stackoverflow.com/questions/35890932/unity-game-manager-script-works-only-one-time/35891919#35891919