Created Triggerable and Trigger scripts.
A triggerable is basically an Event that will call functions to change the scen state.
A trigger is there to handle user input (pointer) and set off one (or more) triggerables.

A triggerable holds several informations regarding whether he is ready to execute.
The triggers should be enabled/disabled according to the position of the player in the scene.