Event entries and callbacks for Triggers are now set up automatically.
Since all the triggers will use the same events and callbacks, it is best ton instantate them automatially at runtime.
However debgging could be trickier, since the callbacks functions added at runtime are not displayed in the EventTrigger inspector.
(The callbacks still work as intended, but this is quite deceiving on first sight)

This will allow to gain some time in defintive since we only have to care about the Triggerers when setting up a Trigger.