Implemented NavGraph builder and creating/updating EdgeTriggers.
This editor class will hold the logic to create or update a NavGraph.
This new class also comes with the ability to automatically create/update the EdgeTriggers in the NavGraph, removing a tedious part of the work to the user.
The EdgeMovement script gets a new GameObject field to hold the EdgeTrigger prefab that UpdateTriggers() needs to instantiate.