Update SplinesOnMove is not useful anymore.
Couldn't fully replicate this Behaviour editor side, so we will keep the file.
The code was commented since the called methods doesn t exist anymore, and their editor equivalent are not accesible.

Furthermore, it caused EdgeTrigger to be instantiated to many times before the building/updating functions were moved in Editor/

The only downside, is it causes a minor regression, we are stil able to automatically update the NavGraph when dragging a Node handle (done in HandleDisplayer script), but not anymore the Node transform.
This could be easily overlooked as we still can use the "update splines" button on the NavGraphManager.