NavGraphBuilder now creates and bind the Paths for the EdgeMovements.
We still have to add intermediate waypoints manually, in the cases we need some.
But since the PathManager already handles this part smoothly, we should leave this as is it is.

At this point the user only need three steps to create a valid NavGraph :
1 - Create the Nodes and place them in the scene
2 - Create the EdgeMovements, and link the Nodes
3 - Click on "Update Splines" to instantiate/place the Paths and EdgeTrigger
Optional - Add intermediate Waypoints whenever relevant
Optional - Manually adjust the EdgesTriggers if needed