Implemented the actual actual gameplay.
The logic of the level is handled in a dedicated manager.
This manager is the hell of a makeshift class, but this idea should get the job done for the moment.
(When we will move on to save and load the state of our levels, we will make a more robust class.)

Added a surface for the exit portal at the end of the level.
This exit is only displayed when the two unlit torches are lit.