Triggers are not anymore disabled from Start.
We may now choose wich ones are enabled at start from the Editor, by disabling their associated collider.
This may seem hackish, but in this game, the colliders only exist for raycasting with the player camera.

This would be a bit more complicate if we needed colliders for actual physics and raycast at the same time.
(We may use the IgnoreRaycast layer in this case, at the expense of collision filtering)