Level_03 : Scripted elevator behavior.
Implemented plateform activation and movement relying upon an Animator.

The player is moved along by adding it as a child element, and replacing is when it was in the scene hierachy when the movement is finished.
(This might not be the cleanest method, but it proved and effective workaround to get things done ASAP)

We also have to keep in mind that the player has to still be on a Node positions when done using the elevator.
An elevator could be seen as a makeshifts, quick and dirty pseudo NavGraph Edge.
This require a bit of extra care when poisitioning the platform and the involved nodes, and creating our animation.
(What would be aweseome was if we could instantiate and set up an Animation and Animator at runtime.)